cmake: child image preload cache value handling
Fix creation of child image preload image cache creation.

The image preload cache file is populated with cache variables.
Unfortunately the value extracted was missing the CACHE type in front.
This means that the value of a local scoped variable of same name as the
cache variable would be used.

Fix this by adding CACHE to ensure it's always the value from the cache
which is placed in the preload file.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>